move: router_category
