+++ app/main.py
@@ -3,6 +3,7 @@ from sqladmin import Admin
 
 from admin import admin_classes
 from config import i_config
+from routers.category_routers import router_category
 from routers.index import router_index
 from models.database import engine
 from routers.client_routers import router_client
@@ -19,6 +20,7 @@ for admin_class in admin_classes:
 app.include_router(router_index)  # , prefix='/api/auth', tags=['auth']
 app.include_router(router)
 app.include_router(router_service)
+app.include_router(router_category)
 app.include_router(router_client)
 if i_config.DEBUG:
     app.include_router(router_init_stub)
