+++ app/routers/service_routers.py
@@ -1,27 +1,14 @@
 from fastapi import Depends, APIRouter, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from models import ServiceNameModel, ServiceCategoryModel
+from models import ServiceNameModel
 from models.database import get_session
-from services.service_service import ServiceNameService, ServiceCategoryService
+from routers.category_routers import ROUTE_CATEGORY
+from services.service_service import ServiceNameService
 
 router_service = APIRouter()
 ROUTE_SERVICE = '/service/'
-ROUTE_CATEGORY = '/category/'
 ROUTE_SERVICE_CATEGORY = f'{ROUTE_SERVICE}{ROUTE_CATEGORY[1:]}'
-
-
-@router_service.get(ROUTE_CATEGORY + '{pk}/', response_model=ServiceCategoryModel)
-async def view_get_service_category_by_id(pk: int, session: AsyncSession = Depends(get_session)):
-    category = await ServiceCategoryService(db_session=session).get(pk=pk)
-    if not category:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'item with id {pk} not found')
-    return category
-
-
-@router_service.get(ROUTE_CATEGORY, response_model=list[ServiceCategoryModel])
-async def view_get_service_category_all(session: AsyncSession = Depends(get_session)):
-    return await ServiceCategoryService(db_session=session).get_all()
 
 
 @router_service.get(ROUTE_SERVICE, response_model=list[ServiceNameModel])
